[NuGet] Clear requireReinstallation in packages.config file on revert.
Changing the project target framework to one that requires the NuGet
packages to be reinstalled and then changing it back again to the
original target framework now removes the requireReinstallation
attribute from the packages.config file. Previously the
requireReinstallation attribute was not removed and the Solution pad
would incorrectly show an error with the package.