[NuGet] Remove updated package available info when update installed.
When the project has updated packages available and an update is
installed then the Packages folder in the Solution window is now
updated to reflect the current state. When the update is installed
the version available information shown next to the package is
removed and the updated package count shown on the Packages folder
is updated.