[Ide] Do not show all files in solution folders.
Addresses bug 19704 - Exception displayed if show all files options is active and user tries to create new folder in newly added solution.

Showing all files in solution files can lead to confusing behaviour when a project is shown as being a child of a solution folder when, in fact, it is not. Solution folders should only show projects that have been explicitly added to that folder.