[NuGet] Fixed bug 20551 - Dependency resolution with custom feeds.
https://bugzilla.xamarin.com/show_bug.cgi?id=20551

A package would fail to install if it depended on another package
that was not in the currently active package source selected in the
Add Packages dialog, but was an enabled package source.

Xamarin Studio now has the same behaviour as Visual Studio. If you
create a NuGet package that depends on Json.NET, store that NuGet
package in a directory on your local machine, you can now define a new
package source pointing to that directory and if the main nuget.org
package source is enabled then you can install the local package into
your project.

The new behaviour is that package dependencies will be resolved from
any enabled package source and not just the package source that is
currently selected in the Add Packages dialog.