[NuGet] Do not check for package updates if package restore fails.
Running package updates after a failed restore will hide the status
bar error message. It also currently causes the Package Console to be
cleared. Also the package update checker would find no packages in the
local solution directory and report that all packages are up to date
which is misleading and potentially incorrect.

The second problem could be worked around by using the information in
the packages.config file (although NuGet's IServiceBasedRepository
GetUpdates method requires a list of IPackage it only uses the
IPackageName, which is the just Id and Version). However unrestored
packages cannot be updated and currently the updated version numbers
are not shown if the packages are missing.