[NuGet] Fix repositories.config file not being restored.
Force the creation of the repositories.config file after the packages
have been restored by creating NuGet Package Managers for each
project.

The repositories.config file is used by NuGet to determine which
projects share the packages directory. All projects, in a solution,
that have NuGet packages should have an entry in the
repositories.config file. If the repositories.config file is missing,
or a project does not have an entry in the file, then NuGet may remove
a NuGet package from the packages directory after uninstalling it from
one project since it believes no other projects in the solution are
using that NuGet package and it can safely be removed completely.