[NuGet] Install all checked packages even if they are not visible.
Previously only those checked packages that were visible in the Add
Packages dialog would be installed when the Add Packages button was
clicked.

Now you can check a package, search for another package, check the new
package, and then install both packages even though the first package
is not displayed in the Add Packages dialog.

A checked package that was previously removed from the Add Packages
dialog by a search will still be checked when it is next displayed
in the Add Packages dialog if the dialog is not closed. So if the
search is cleared or the search returns the original package it will
show as checked. Note that the package id and the package version
needs to exactly match. A different version of the package would not
be checked.