[NuGet] Uncheck original package when a different version checked.
Previously it was possible to check multiple different versions of the
same package. This would mean the Add Packages dialog would install a
package multiple times and the end result would be just one package
installed. The last package to be installed would be the one left in
the project. The other versions would have been uninstalled as each
package was installed.

Now when a different version of that package is checked the previously
checked package is unchecked. So only one package with the same id
will be installed from the Add Packages dialog. This also means when a
search is run for all package versions it is not possible to check
multiple packages when they have the same package id. Since it is not
possible to install multiple versions of the same package into the
same project it should not be possible to do this from the Add
Packages dialog.