[Core] Fix project reloading issue
The ReloadItem method now takes into account that the Items list may have
changed while loading the new instance of the project. It may happen for
example if the project is migrated and that causes a file save, and that
causes a reload.

This fixes bug 20615 - XS throws exception on migrating MonoMac project
to XamMac through Projects-> Migrate to Xamarin.Mac