[NuGet] Show packages up to date message in status bar.
When the user updates a single package, all packages in the project or
all packages in the solution and all packages are already up to date
then the status bar now displays a message indicating that the package
or packages are up to date.

Previously the NuGet addin would display a message that the package or
packages were successfully updated when nothing actually happened.