[NuGet] Fix version shown as download count.
Fixed bug 21440:  [Add Package Dialog] In the search package field,
when entering the text as 'AutoMapper version:*' then in the package
details section the download shows the package version.
https://bugzilla.xamarin.com/show_bug.cgi?id=21440

When running a package version search (PackageId version:Number) the
right hand side of the dialog was showing the version number instead
of the download count.

Now it shows the download count. Ideally it would show the download
count of that particular version but this is not currently available
from the Package object. It is returned in the results back from the
package source but it is not available on the Package object. This
probably needs a change to NuGet so the value can be accessed.

Also the download counts are different for the same package when
doing a normal search or a package version search. The package version
search has a larger download count number. Not sure if this related to
the stats problem NuGet had recently:

http://blog.nuget.org/20140603/nuget-stats.html

Currently the NuGet addin is showing the download count it receives.
The standard search download counts match those shown in Visual
Studio's Manage Packages dialog. For the package version search the
download count value matches that shown in the NuGet.org website for
an individual package (e.g. https://www.nuget.org/packages/jQuery).