[NuGet] Fix updating packages marks packages.config as deleted by Git
Fixed bug 21327 - Updating NuGet packages can cause packages.config to
be treated as deleted by git
https://bugzilla.xamarin.com/show_bug.cgi?id=21327

On updating packages in a project, and the project is using Git, then the
update can cause Git to treat the packages.config file as deleted.

1) Create a C# library project.
2) Add Json.NET version 5.0.8 NuGet package to the project. (use the NuGet
package version search "newtonsoft.json version:*" so you can pick an older
version of the NuGet package.
3) Add the project to a local Git repository (git init)
4) Then commmit the project files into the local git repository.
5) Close and re-open the solution.
6) Now right click the Packages folder and select Update.
7) Using git from the command line use gitk to see what happened in Git.
8) packages.config file is marked as deleted in the index.

The packages.config file should be shown as modified but not deleted.
The packages.config file would not be deleted if there were other
NuGet packages in the file that were not affected by the update. It is
deleted because the old package is first uninstalled, NuGet then sees
there are no NuGet packages referenced and deletes the packages.config
file.

The NuGet addin now does not use the MonoDevelop.Core.FileService's
DeleteFile method if the packages.config file is deleted when an
package update occurs. Instead the NuGet addin deletes the file
itself. NuGet.Core will recreate the packages.config file when the
new package version is installed. This will mean the packages.config
file will show as modified by Git and not deleted.